tests/wpt/tests/css/css-align/parsing/justify-content-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: getComputedStyle().justifyContent</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-content">
<meta name="assert" content="justify-content computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("justify-content", "normal");

test_computed_value("justify-content", "space-between");
test_computed_value("justify-content", "space-around");
test_computed_value("justify-content", "space-evenly");
test_computed_value("justify-content", "stretch");

test_computed_value("justify-content", "center");
test_computed_value("justify-content", "start");
test_computed_value("justify-content", "end");
test_computed_value("justify-content", "flex-start");
test_computed_value("justify-content", "flex-end");
test_computed_value("justify-content", "unsafe end");
test_computed_value("justify-content", "safe flex-start");
test_computed_value("justify-content", "left");
test_computed_value("justify-content", "unsafe right");
</script>
</body>
</html>
